% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airport_data}
\alias{airport_data}
\title{Origin Airport IDs and their respective Longitude and Latitude coordinates}
\format{
a dataframe with 354 rows and 3 columns
\describe{
   \item{Origin}{airport city code}
   \item{Longitude}{longitude of airport, in decimal degrees east}
   \item{Latitude}{latitude of arport, in decimal degrees north}
}
}
\usage{
airport_data
}
\description{
A dataset with the coordinates of 354 airports
}
\keyword{datasets}
