% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_symmetry.R
\name{check_symmetry}
\alias{check_symmetry}
\title{Checks whether a list is symmetrical}
\usage{
check_symmetry(direct_from)
}
\arguments{
\item{a}{list, ex(direct_from or sym_direct_from)}
}
\value{
a list whose first element is TRUE or FALSE (whether or not it is symmetric), and whose second element is an `n x 2` character matrix giving the `n` asymmetric connections.
}
\description{
Tests whether a `direct_from` object is symmetric.
}
\examples{
test1 <- check_symmetry(direct_from_a)
print(test1)
test2 <- check_symmetry(sym_direct_from)
print(test2)
}
