% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{closest_airport100}
\alias{closest_airport100}
\title{Airports within 100 miles of a Longitude and Latitude coordinate}
\format{
A list with 9285 element. Each element is indexed by an indices in airport_data.
\describe{
   \item{Point}{lon lat}
   \item{Airports}{char vector of connecting flights to airports}
}
}
\usage{
closest_airport100
}
\description{
Airports within 100 miles of a Longitude and Latitude coordinate
}
\keyword{datasets}
