% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{direct_from_a}
\alias{direct_from_a}
\title{Direct_from object for scenario A: The total number of direct connections must remain constant. In other words, if you add a connection, you must remove a different connection.}
\format{
a list with 354 elements
\describe{
   \item{Airport}{airport city code}
   \item{Connectinng flights}{char vector of connecting flights to airports}
}
}
\usage{
direct_from_a
}
\description{
A list with 354 elements, each element is a char vector of airports with direct flights from element airport
}
\keyword{datasets}
