% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{direct_from_b}
\alias{direct_from_b}
\title{Direct_from object for scenario B: The total number of connections to and from each airport must remain constant.}
\format{
a list with 354 elements
\describe{
   \item{Airport}{airport city code}
   \item{Connectinng flights}{char vector of connecting flights to airports}
}
}
\usage{
direct_from_b
}
\description{
A list with 354 elements, each element is a char vector of airports with direct flights from element airport
}
\keyword{datasets}
