% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{direct_from_c}
\alias{direct_from_c}
\title{Direct_from object for scenario C: add or subtract up to two connections from each airport, leaving the other connections unchanged, and keeping the total number of connections constant.}
\format{
a list with 354 elements
\describe{
   \item{Airport}{airport city code}
   \item{Connectinng flights}{char vector of connecting flights to airports}
}
}
\usage{
direct_from_c
}
\description{
A list with 354 elements, each element is a char vector of airports with direct flights from element airport
}
\keyword{datasets}
