% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distfun.R
\name{distfun}
\alias{distfun}
\title{Calculates the Great Circle Distance}
\usage{
distfun(lonlat1, lonlat2)
}
\arguments{
\item{lonlat1, }{an n2 x 2 matrix, first column longitudes(positive for eastern hemisphere and negative for western hemisphere), second column latitude (positive for northern hemisphere and negative for southern hemisphere)}
}
\value{
distmat, an n1 x n2 matrix of the pairs' distance (ie.distmat[i,j] is the distance between lonlat1[i,] and lonlat2[j,])
}
\description{
Computes the Great Circle Distance distance from two coordinates between
pairs of longitude and latitude coordinates
}
\examples{
# Calculate the distance between all the airports in the airport_data dataframe
lonlat1 <- data.frame(airport_data$Longitude, airport_data$Latitude)
lonlat2 <- data.frame(airport_data$Longitude, airport_data$Latitude)
test <- distfun(lonlat1, lonlat2)
print(head(test) )
}
