% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lonlat_usa}
\alias{lonlat_usa}
\title{Grid of USA Longitude and Latitude coordinates}
\format{
a dataframe with 9284 rows and two columns
\describe{
   \item{Longitude}{longitude of airport, in decimal degrees east}
   \item{Latitude}{latitude of arport, in decimal degrees north}
}
}
\usage{
lonlat_usa
}
\description{
A dataset with the coordinates of coordinates in USA
}
\keyword{datasets}
