% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_near_airport75}
\alias{points_near_airport75}
\title{Airports within 75 miles of a Longitude and Latitude coordinate}
\format{
A list with 9285 element. Each element is indexed by an indices in airport_data.
\describe{
   \item{Airport}{airport city code}
   \item{Connectinng flights}{char vector of connecting flights to airports}
}
}
\usage{
points_near_airport75
}
\description{
Airports within 75 miles of a Longitude and Latitude coordinate
}
\keyword{datasets}
