% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_near_points100}
\alias{points_near_points100}
\title{Points within 100 miles of each other}
\format{
A list with 9285 element. Each element is indexed by an indices in airport_data.
\describe{
   \item{Points}{lonlat}
   \item{naerby points}{list of lonlats }
}
}
\usage{
points_near_points100
}
\description{
Points within 100 miles of each other
}
\keyword{datasets}
