% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points_near_points75}
\alias{points_near_points75}
\title{Points within 75 miles of each other}
\format{
A list with 9285 element. Each element is indexed by an indices in airport_data.
\describe{
   \item{Point}{airport city code}
   \item{Point}{lonlat of nearby points}
}
}
\usage{
points_near_points75
}
\description{
Points within 75 miles of each other
}
\keyword{datasets}
