% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sym_direct_from}
\alias{sym_direct_from}
\title{Symmetric direct from list of airport connects}
\format{
a list with 354 elements
\describe{
   \item{Airport}{airport city code}
   \item{Connectinng flights}{char vector of connecting flights to airports}
}
}
\usage{
sym_direct_from
}
\description{
A list with 354 elements, each element is a char vector of airports with direct flights from element airport
}
\keyword{datasets}
